fix: Remove unnecessary state persistence in createMetaRPCHandler (#24851)
https://github.com/MetaMask/metamask-extension/pull/15978 attempted to
deal with the following problem:

"in MV3 the background process (now a service worker) is constantly (and
unpredictably) winding down... there is a small window of time when, if
the service worker's lifecycle ends immediately after a user action and
the action is executed in the background but not yet persisted, the
action will not be added to the replay queue since the action was in
fact executed - just not yet persisted."

The solution was: "This solution we've landed on is to persist the data
that results from background state changes further upstream than
previously, in the controller multiplex. This allows us to persist the
data before the action is resolved ensuring that the queuing system will
work "

The problem with the solution is that the metaRPChandler would now have
to do a write to store after every single request, and if multiple of
these are occuring around the same time, they end up blocking subsequent
requests (as currently storage writes can take upwards of 100 ms on some
systems)

Fortunately, the original problem is no longer a problem, as the MV3
service worker will no longer be stopped if the UI is open. So, we can
delete the blocking code, as it is no longer needed.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/24851?quickstart=1)

Fixes: https://github.com/MetaMask/metamask-extension/issues/24583

On a linux machine:

1. Go to the test dapp
2. Trigger 10 batx tx
3. Accept/Reject multiple times in a row
4. It should be as performant as MV2

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>